refactor: manual relationship handling
