+++ src/common/DbTables/BoardTable.ts
@@ -1,4 +1,4 @@
-import Board from '../../resources/boards/board.model';
+import Board from '../../resources/boards/board.entity';
 import Table from './Table';
 
 class BoardTable extends Table<Board> {}
